fix: 레퍼런스 좌우 비대칭의 진짜 원인 세 가지 — float32 캐스팅이 주범
v35 이후 모든 정책이 물려받은 레퍼런스 비대칭(hip_pitch +41%)의 정체를 찾았다.
placo 원본 녹화는 +-2% 로 대칭인데 파이프라인이 깨뜨리고 있었다.

1) fit_poly.py 가 시간축을 float32 로 캐스팅했다. 15차를 t in [0,1) 에서
   맞추는 것은 조건수가 나빠(RankWarning 수백 건) 입력 정밀도가 그대로
   결과에 나온다. 같은 데이터·같은 차수에서
       float32 X : 재현오차 6.883도, 무릎 좌우차 +16.4%
       float64 X : 재현오차 0.786도, 무릎 좌우차  -3.1%  (녹화 -1.8%)

2) poly_reference_motion.py 가 계수를 float32 텐서에 담았다. float64 적합은
   계수가 1e8 대라 float32 로는 Horner 평가의 자릿수 소거를 못 버틴다
   (무릎 좌우차가 +176% 까지 튐). 평가를 float64 로, 반환만 float32 로.

3) auto_waddle.py 속도 필터가 격자의 45% 를 삭제했다 (medium 대역 밖 녹화를
   지우는데 스윕은 0.222 까지 돈다). --no-speed-filter 추가.

부수: 적합 시간축이 런타임과 달랐다 (linspace(0,1,N) 간격 1/(N-1) vs 런타임
1/N). 직진 녹화는 반주기 거울로 시계열에서 대칭화한다
(scripts/setup/symmetrize_recordings.py) — 적합 후 대칭화는 위상 감김
불연속 때문에 폭주해서 버렸다.

결과 ref_g125sym.pkl: 격자 충전율 55%->100%, dy=0 확보,
hip_pitch +41.0%->-2.9%, knee +11.1%->+2.5%, ankle -8.0%->+1.2%.

+++ source/open_duck_mini_isaaclab/reference_motion/poly_reference_motion.py
@@ -114,7 +114,13 @@ class PolyReferenceMotion:
             )
             return parsed[best[0]][best[1]][best[2]]
 
-        grid = torch.zeros(nb_dx, nb_dy, nb_dtheta, REF_FRAME_DIM, degree_plus_1, device=self.device)
+        # **float64 여야 한다.** 15 차 다항식을 t∈[0,1) 에서 맞추면 계수가 1e8 대까지
+        # 커지고 Horner 평가에서 대규모 자릿수 소거가 일어난다. float32(유효숫자 7)로는
+        # 그 소거를 못 버텨서 결과가 무너진다 — 2026-08-11 실측으로 무릎 진폭 좌우차가
+        # -3 % 여야 할 것이 +176 % 까지 튀었다. 표 자체는 작고(수십만 원소) 평가도
+        # [N,36,16] 뿐이라 float64 비용은 무시할 만하다. 반환할 때 float32 로 되돌린다.
+        grid = torch.zeros(nb_dx, nb_dy, nb_dtheta, REF_FRAME_DIM, degree_plus_1,
+                           device=self.device, dtype=torch.float64)
         for xi, dx in enumerate(self.dxs):
             for yi, dy in enumerate(self.dys):
                 for ti, dtheta in enumerate(self.dthetas):
@@ -122,7 +128,7 @@ class PolyReferenceMotion:
                         coeffs = parsed[dx][dy][dtheta]  # list of D lists, each length K
                     else:
                         coeffs = nearest_coeffs(dx, dy, dtheta)
-                    grid[xi, yi, ti] = torch.tensor(coeffs, device=self.device)
+                    grid[xi, yi, ti] = torch.tensor(coeffs, device=self.device, dtype=torch.float64)
 
         self.coeffs = grid
         self.dxs_t = torch.tensor(self.dxs, device=self.device)
@@ -149,11 +155,11 @@ class PolyReferenceMotion:
 
         env_coeffs = self.coeffs[ix, iy, ith]  # [N, 36, K]
 
-        t = (i.float() % self.nb_steps_in_period) / self.nb_steps_in_period
-        t = t.clamp(0.0, 1.0)  # [N]
+        t = (i.double() % self.nb_steps_in_period) / self.nb_steps_in_period
+        t = t.clamp(0.0, 1.0)  # [N], float64 — 위 grid 주석 참고
 
         num_envs, dim, degree_plus_1 = env_coeffs.shape
-        result = torch.zeros(num_envs, dim, device=self.device)
+        result = torch.zeros(num_envs, dim, device=self.device, dtype=torch.float64)
         for k in range(degree_plus_1):
             result = result * t.unsqueeze(-1) + env_coeffs[:, :, k]
-        return result
+        return result.float()
